pyproject.toml
[project]
name = "simpsons"
version = "0.1.0"
description = "Little silly project"
authors = [
    {name = "Renata Kostolina",email = "kostolina.renata@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch (>=2.6.0,<3.0.0)",
    "sphinx (>=4.0,<5.0)",
    "autodocsumm (>=0.2.14,<0.3.0)",
    "pytorch-lightning (>=2.5.1,<3.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "dvc (>=3.59.1,<4.0.0)",
    "pytest-cov (>=6.1.1,<7.0.0)",
    "pytest-regressions (>=2.7.0,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
]


[tool.black]
line-length = 100
[tool.flake8]
max-line-length = 100
ignore = ['B018', 'F841', 'W503']


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
